# docs/api-docs.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/v1/events/{id}":{"get":{"tags":["Events Management"],"summary":"Get my event with id","description":"Endpoint to get all events for the logged in user","operationId":"getEventById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Events retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventDetailPageResponse"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"example":{"code":"unauthorized","message":"You are not authorized to access this resource"}}}}},"security":[{"Authorization":[]}]},"put":{"tags":["Events Management"],"summary":"Update an event","description":"Endpoint to update an event. You can update only your events.","operationId":"putEvents","parameters":[{"name":"id","in":"path","description":"The ID of the event to be retrieved","required":true,"schema":{"type":"string"}}],"requestBody":{"description":"The event to update","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventCrudRequest"}}},"required":true},"responses":{"200":{"description":"Event updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventLightResponse"}}}},"400":{"description":"Bad Request","content":{"application/json":{"example":{"code":"bad_request","message":"Event name is too large"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"example":{"code":"unauthorized","message":"You are not authorized to access this resource"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"example":{"code":"internal_server_error","message":"Internal Server Error"}}}}},"security":[{"Authorization":[]}]},"delete":{"tags":["Events Management"],"summary":"Delete an event","description":"Endpoint to delete an event. You have to be logged as ADMIN to delete an event.","operationId":"deleteEvent","parameters":[{"name":"id","in":"path","description":"The ID of the event to be retrieved","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"Event deleted successfully"},"400":{"description":"Bad Request"},"401":{"description":"Unauthorized"},"500":{"description":"Internal Server Error"}},"security":[{"Authorization":[]}]}},"/auth/signup":{"post":{"tags":["Authentication"],"summary":"User Signup","description":"Endpoint to register a new user","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSignupRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}}}},"/auth/login":{"post":{"tags":["Authentication"],"summary":"User Login","description":"Endpoint to authenticate a user","operationId":"authenticate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}}}},"/api/v1/organizer":{"post":{"tags":["User Management"],"summary":"Create your organizer data","description":"Endpoint to create an organizer data for a user","operationId":"createOrganizer","requestBody":{"description":"The organizer data to create","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrganizerCrudRequest"}}},"required":true},"responses":{"201":{"description":"Organizer created successfully"},"400":{"description":"Bad Request"},"401":{"description":"Unauthorized"},"500":{"description":"Internal Server Error"}},"security":[{"Authorization":[]}]}},"/api/v1/events":{"get":{"tags":["Events Management"],"summary":"Get my events","description":"Endpoint to get all events for the logged in user","operationId":"getEvents","responses":{"200":{"description":"Events retrieved successfully","content":{"application/json":{}}},"401":{"description":"Unauthorized","content":{"application/json":{"example":{"code":"unauthorized","message":"You are not authorized to access this resource"}}}}},"security":[{"Authorization":[]}]},"post":{"tags":["Events Management"],"summary":"Create a new event","description":"Endpoint to create a new event. You have to be logged as ADMIN to create an event.","operationId":"postEvents","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventCrudRequest"}}},"required":true},"responses":{"201":{"description":"Event created successfully"},"400":{"description":"Event is not valid","content":{"application/json":{"example":{"code":"bad_error","message":"Event name is too large"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"example":{"code":"unauthorized","message":"You are not authorized to access this resource"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"example":{"code":"internal_server_error","message":"Internal Server Error"}}}}},"security":[{"Authorization":[]}]}},"/users":{"get":{"tags":["User Management"],"summary":"Get all users","description":"Endpoint to retrieve all users in the system","operationId":"allUsers","responses":{"200":{"description":"Users retrieved successfully","content":{"application/json":{}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}}}},"security":[{"Authorization":[]}]}},"/users/me":{"get":{"tags":["User Management"],"summary":"Get authenticated user","description":"Endpoint to retrieve the currently authenticated user","operationId":"authenticatedUser","responses":{"200":{"description":"User retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"Authorization":[]}]}},"/api/public/v1/event/{id}":{"get":{"tags":["Public Endpoints"],"summary":"Get all the information of an event by ID","description":"Endpoint to get a specific event by its ID formatted for detail page","operationId":"getEvent","parameters":[{"name":"id","in":"path","description":"The ID of the event to be retrieved","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Event retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventDetailPageResponse"}}}},"400":{"description":"Event id is not valid","content":{"application/json":{"example":{"code":"bad_found","message":"Event id is not valid"}}}},"404":{"description":"Event not found","content":{"application/json":{"example":{"code":"not_found","message":"Event not found"}}}}}}}},"components":{"schemas":{"EventCrudRequest":{"type":"object","properties":{"title":{"type":"string","description":"The title of the event","example":"Concierto de Trueno"},"date":{"type":"string","format":"date-time","description":"The date of the event","example":"2022-01-01T20:00:00"},"location":{"$ref":"#/components/schemas/LocationDTO","description":"The location of the event"},"image":{"$ref":"#/components/schemas/ImageDTO","description":"The image of the event"},"prices":{"type":"array","description":"The prices of the event","items":{"$ref":"#/components/schemas/PriceDTO"}},"description":{"type":"string","description":"The description of the event","examples":["Trueno es uno de los grandes artistas latinoamericanos de nuestros tiempos. Un rapero que comenzó destacándose en las batallas de freestyle en su natal Buenos Aires y hoy ha logrado el reconocimiento internacional además de acumular millones de oyentes en plataformas digitales. Este 7 de junio el Movistar Arena será testigo del prime de uno de los nuevos grandes del universo urbano."]},"additionalInfo":{"type":"array","description":"The additional info of the event","example":["Prohibido el ingreso de menores de edad","No se permite el ingreso de alimentos y bebidas","No se permite el ingreso de cámaras fotográficas"],"items":{"type":"string"}}}},"ImageDTO":{"type":"object","properties":{"url":{"type":"string","description":"The URL of the image","example":"https://movistararena.co/wp-content/uploads/2025/02/trueno-2025-4.jpg"},"alt":{"type":"string","description":"The alt text of the image","example":"Concierto de Trueno"}}},"LocationDTO":{"type":"object","properties":{"name":{"type":"string","description":"The name of the location","example":"Movistar Arena"},"address":{"type":"string","description":"The address of the location","example":"Dg. 61c #26-36"},"city":{"type":"string","description":"The city of the location","example":"Bogotá"}}},"PriceDTO":{"type":"object","properties":{"id":{"type":"string","description":"The ID of the price","example":"001b2f30-9a84-45e1-9345-518bea8a77c8"},"value":{"type":"number","format":"double","description":"The value of the price","example":100.0},"currency":{"type":"string","description":"The currency of the price","example":"$"},"type":{"type":"string","description":"The type of the price","example":"General"},"isFree":{"type":"boolean","description":"The description of the price","example":false}}},"EventLightResponse":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"date":{"type":"string","format":"date-time"},"location":{"type":"string"},"bannerUrl":{"type":"string"},"price":{"type":"number","format":"double"},"currency":{"type":"string"},"status":{"type":"string"}}},"UserSignupRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"email":{"type":"string"}}},"LoginResponse":{"type":"object","properties":{"token":{"type":"string"},"expiresIn":{"type":"integer","format":"int64"}}},"UserLoginRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"OrganizerCrudRequest":{"type":"object","properties":{"name":{"type":"string","description":"The name of the organizer","example":"Movistar Arena"},"url":{"type":"string","description":"The description of the organizer","example":"https://movistararena.co/"},"logo":{"$ref":"#/components/schemas/ImageDTO","description":"The logo of the organizer"}}},"OrganizerDTO":{"type":"object","properties":{"id":{"type":"string","description":"The ID of the organizer","example":"001b2f30-9a84-45e1-9345-518bea8a77c8"},"name":{"type":"string","description":"The name of the organizer","example":"Movistar Arena"},"url":{"type":"string","description":"The description of the organizer","example":"https://movistararena.co/"},"logo":{"$ref":"#/components/schemas/ImageDTO","description":"The logo of the organizer"}}},"UserResponse":{"type":"object","properties":{"id":{"type":"string"},"username":{"type":"string"},"email":{"type":"string"},"role":{"type":"array","items":{"type":"string"}},"organizer":{"$ref":"#/components/schemas/OrganizerDTO"}}},"EventDetailPageResponse":{"type":"object","properties":{"id":{"type":"string","description":"The ID of the event","example":"cbb46e0f-1014-41f9-bf63-ba1203197ce8"},"title":{"type":"string","description":"The title of the event","example":"Concierto de Trueno"},"date":{"type":"string","format":"date-time","description":"The date of the event","example":"2022-01-01T20:00:00"},"location":{"$ref":"#/components/schemas/LocationDTO","description":"The location of the event"},"image":{"$ref":"#/components/schemas/ImageDTO","description":"The image of the event"},"prices":{"type":"array","description":"The prices of the event","items":{"$ref":"#/components/schemas/PriceDTO"}},"description":{"type":"string","description":"The description of the event","examples":["Trueno es uno de los grandes artistas latinoamericanos de nuestros tiempos. Un rapero que comenzó destacándose en las batallas de freestyle en su natal Buenos Aires y hoy ha logrado el reconocimiento internacional además de acumular millones de oyentes en plataformas digitales. Este 7 de junio el Movistar Arena será testigo del prime de uno de los nuevos grandes del universo urbano."]},"additionalInfo":{"type":"array","description":"The additional info of the event","example":["Prohibido el ingreso de menores de edad","No se permite el ingreso de alimentos y bebidas","No se permite el ingreso de cámaras fotográficas"],"items":{"type":"string"}},"organizer":{"$ref":"#/components/schemas/OrganizerDTO","description":"The organizer of the event information"},"status":{"type":"string","description":"The status of the event","example":"ACTIVE"}}}},"securitySchemes":{"Authorization":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}